+++ src/core/server/commands/revive.ts
@@ -6,8 +6,7 @@ addCommand('revive', handleCommand);
 
 function handleCommand(player: alt.Player, targetPlayerID: string | null = null): void {
     if (targetPlayerID === null) {
-        player.send(CommandsLocale.CANNOT_WHILE_DEAD);
-        finishRevive(player);
+        player.handleDeathRespawn(player.pos);
         return;
     }
 
@@ -22,14 +21,5 @@ function handleCommand(player: alt.Player, targetPlayerID: string | null = null)
         return;
     }
 
-    finishRevive(target);
-}
-
-function finishRevive(target: alt.Player) {
-    target.spawn(target.pos.x, target.pos.y, target.pos.z, 0);
-    target.safeAddHealth(200, true);
-    target.safeAddArmour(0, true);
-    target.data.isDead = false;
-    target.saveField('isDead', false);
-    target.nextDeathSpawn = null;
+    target.handleDeathRespawn(target.pos);
 }